requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
awsebcli==3.14.7
blessed==1.15.0
botocore==1.12.5
cached-property==1.4.3
cement==2.8.2
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3
cryptography-vectors==2.3
Cython==0.28.3
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-progressive-web-app==0.1.1
django-rest-auth==0.9.3
django-rest-knox==3.1.5
django-rest-swagger==2.1.2
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
docker==3.5.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
environs==4.0.0
Flask==1.0.2
Flask-RESTful==0.3.6
flask-restplus==0.11.0
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.6
isort==4.3.4
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
marshmallow==2.15.4
mccabe==0.6.1
more-itertools==4.3.0
nltk==3.3
nose==1.3.7
numpy==1.14.5
oauthlib==2.1.0
openapi-codec==1.3.2
pathlib==1.0.1
pathspec==0.5.5
Pillow==5.2.0
pluggy==0.8.0
psycopg2==2.7.5
py==1.7.0
pyasn1==0.4.2
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.6.4
pylint==2.1.1
pyOpenSSL==18.0.0
PySocks==1.6.8
pytest==3.10.0
python-dateutil==2.6.1
python-dotenv==0.9.1
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.13
requests==2.18.4
requests-oauthlib==1.0.0
rsa==3.4.2
s3transfer==0.1.12
semantic-version==2.5.0
simplejson==3.16.0
six==1.11.0
SQLAlchemy==1.2.8
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.22
waitress==1.1.0
wcwidth==0.1.7
websocket-client==0.51.0
Werkzeug==0.14.1
whitenoise==3.3.1
wrapt==1.10.11
